vsRenderQueueLayer renamed to vsRenderQueueStage
To remove confusion over the "layer" which can be set on a material.
"Stage" can be normal, glow, or post-glow.  "layer" sorts rendering
within those rough categories.

"postglow" can now be set on a material, to force it to be rendered
after glowing objects.  (Almost never what you actually want, but I have
need of it for a specific effect)