Avoids rebuilding OpenGL context on nop resize events
Previously, we'd rebuild our OpenGL rendering context if we received a
resize event which didn't actually change any of our rendering settings.
(This happened under OS X if we minimised and then restored a
VectorStorm application).  Now, we don't rebuild the context unless the
size, bitdepth, or fullscreen flag changes.