Added vsRenderBuffer::P type.
Pure vertex position data.  Enough stuff in VectorStorm's core codebase
was starting to rely on being able to read the data in a vsRenderBuffer
that it was getting a little silly for there not to be support for
pure position data.  A lot of bounding box utility functions, in
particular, used to not support pure vertex data entered via the generic
interface to vsRenderBuffer, so they couldn't tell what type of data
was in there.