Fixes render sorting edge case error.
When there was something which should have been the first thing to
draw in a scene (but which wasn't the first thing _submitted_ to be
drawn in the scene), it would previously have been sorted into drawing
second.  In most cases, this made no difference (since there are usually
lots of items which are all set up to draw at the same priority level),
but this could lead to bad behaviour in simple scenes.