Adds support for setting glViewport for a vsScene.
vsScene::SetViewport() sets the portion of the screen which should be
used for rendering that scene.  This takes a Box2D specifying the
region of the screen which will be drawn to.  X and y coordinates are
expected to be expressed in the range of [0..1] for left..right and
bottom..top.