Fixed non-resize in automatic vsArrayStore resize.
Would result in buffer overflow and other badness.  Am chagrined and
much embarrassed that (a) this major, stupid bug made it into the
codebase, (b) has been here for so long, (c) that I implemented an
automatic resize in these 'array' classes to begin with.  Why don't I
just use STL if I'm going to do that sort of performance-hostile thing?

Need to consider rolling back this feature;  arrays should be
initialised to their capacity, and NOT ever auto-resize -- trying to
over-fill one of them should be a fatal error in game code.  (Note that
this opinion of mine is probably why I never noticed that the
auto-resize was there, and that it was so horribly broken)