Proper invert for 3x3 matrix.
Previously, the vsMatrix3x3::Invert() and Inverse() functions actually
performed a transpose (since that's faster to calculate for rotation
matrices, which is what you usually put in a vsMatrix3x3).  However, it
wasn't a real, generic invert.

I've now implemented a proper invert, and moved the old, fast "basically
the same as an invert for rotation matrices" over to 'Transpose()'.
Nothing inside VectorStorm was calling that function anyway, and I only
had one spot in a game which was calling it (and even that was old code
which was no longer in use).  Hopefully this shouldn't affect anyone.