Support for resizing game window.  Not yet exposed.
This adds the necessary infrastructure to vsRenderer and vsRenderScheme
to support resizing a game window while the game is running.  Note that
games cannot currently enable this;  I need to work out how a game
should specify to the VectorStorm engine that it wants to support
live resizing.  But apart from that, it's now working.

Note that games which support resizing might require some extra work.
There is currently no notification to the game that a resize has
happened, which means that 2D games which position elements on screen
based on the calculated "corners" of the screen need to constantly be
recalculating where things go on the screen horizontally (since aspect
ratio affects the horizontal position of the corners).

Really, I need a mechanism for the engine to inform the game that the
resolution has changed.  Some sort of callback or event handling code,
perhaps.  Or maybe build it into the coreGame class, and just have that
poll the engine every frame, checking for resolution changes and then
dispatching an event into client game code.

Still a lot of architectural work to be done on this, to determine
exactly how to fit this into the overall engine structure.  But at least
the mechanics of resizing the window are now working, even if I'm not
quite sure how to structure code to make it fit with the rest of the
engine, yet.