Adds support for scissoring.
Like viewports, scissoring is defined using a box in the range [0..1] on
x and y, which the engine will convert to be relative to the game's
resolution.

When scissoring has been turned on, all subsequent drawing operations
will be clipped to within the specified box.  It's basically a cheaper,
simpler version of stenciling.

Note that you'll need to be doing something special to make much use of
scissoring -- if you're rendering using the vsRenderQueue, your
rendering commands will be reordered by the queue in order to minimise
material changes, so you almost certainly won't wind up with the results
you were expecting.  Easiest way to use scissoring successfully is to
perform scissored rendering directly, by issuing drawing commands into
the vsRenderQueue's generic display list, which won't try to optimise
your drawing instructions for you.