Experimental support for boxing text strings.
(That is, specifying a "box" inside which the final, formatted text must
fit.  The font renderer will shrink the font as necessary in order to
force the text to fit inside the specified box)

Note that only the box's dimensions are checked;  not its absolute
positioning.

Also note that this has (so far) only been implemented for the basic
rendering of fonts into fragments, not any of the older interfaces, or
the special-case interfaces which build strings in fancy ways.