Default material 'blend' to true.
Was previously defaulting to false.  But nobody was ever setting it, and
the renderer was defaulting it to true.  Net result:  Set 'blend' to
anything on even one material, and suddenly it turns off for everything
else.

In practice, we usually want 'blend' enabled.  Can still turn it off
when we actually want to get rid of it.