vsFont now handles multiple text sizes, display density.
General strategy:  A file can now be created which specifies several
".fnt" files.  When a game asks to draw in a particular size, the
vsFont compares the requested text size to the provided font sizes, and
will choose the *equal or larger* font size to which it has access, and
will perform the drawing with that font.

Additionally, the font keeps track of all fragments it has created.
When desired (for example, when the window is moved from a display with
one pixel density and to one with a different density), the game may ask
the font to rebuild all of its still-existing fragments.  This allows us
to seamlessly drag the game window from a Retina display to a non-Retina
display on OSX, and retain pixel-perfect font rendering on both sides.

Downside:  this requires advance loading of high resolution versions of
fonts, even if the computer being used doesn't support highdpi displays.