Adds a GetClosestTimeTo() utility function to vsSpline3D.
Refactors GetClosestPointTo() to use it.  This is just because many uses
of this function might want the velocity or the time, in addition to the
point which is closest to the passed time.  By returning the spline time
of the closest point, callers may fetch whatever spline information they
want.