Fixes MouseMotion input on 2D scenes.
Previously, camera position in 2D scenes would be applied to
'MouseMotion' input, as if that camera position was part of the mouse
motion every frame.  Now the mouse motion is only adjusted for the
screen's visible scale, not for the camera offset.

(TODO:  Consider;  should we use the camera rotation as well?)