Rename CID_ExitApplication to CID_Escape.
ExitApplication is an operation which may or may not be enabled for any
particular game.  (And if disabled, then 'ExitApplication' is definitely
the wrong appellation for whatever the game is doing with the button)

Instead, let's name the button 'Escape' (as that's the default mapping,
on a keyboard, but could be mapped to whatever other physical button is
desired).

'CID_Exit' should also be reconsidered.