Fix creation of vsQuaternions with angle values outside of [PI..-PI].
Note that this fix isn't actually correct;  quaternions can represent
any rotation, and this fix clamps them into [-PI .. PI], by using a
vsAngle to wrap them into that range.

The previous code was extremely wrong if values outside that range were
used, so this fix is an improvement;  it's just not ideal, yet.