Set 3D line widths separately per side.
"left" and "right" sides, from the point of view of the line being
drawn, relative to the camera's facing.

So for example, if the line is being drawn from the left side of the
screen to the right side, the "left" width is toward the top of the
screen, while the "right" width is toward the bottom.