vsImage now uses a raw data format, internally.
This is much, much better for memory usage (1/4 the memory use), and
also uses far less memory and incurs far less conversion cost when
converting a vsImage into a vsTexture.

Also adds a 'Blit()' method on vsTextureInternal, which you may use to
blit a vsImage into a texture.  This is going to be the standard
interface for partial texture updates.

Note that the *whole* vsImage gets blitted into the vsTextureInternal,
using an offset.  So if the idea is to update just a portion of the
vsTextureInternal, you don't want to be using the same vsImage that
originally created it;  you want to be using a smaller vsImage
containing only the updated data.