Adjust data directory on Win32.
While in development, the exe is commonly located in a 'Debug' or 'Release'
directory hanging off the project directory, and the data is never copied into
there.  This code checks whether that's the case, and if so, tells PhysFS to
look for the Data directory one level outside the executable's location.