VectorStorm now drops its framerate when its window is hidden.
Previously, did it when the window lost focus.

Also, added log messages for the other SDL window events (which look
like they duplicate the functionality of other events?  But at least
this will help me figure out exactly when which events are sent)

TODO:  Should this be configurable, instead of hardcoded behaviour?