Adds new font justification options.
Can now justify fonts to Top and Bottom, in addition to Left and Right
(and all such combinations).  When justified to Top, the font draws so
that the TOP of the rendered text appears at the specified position.
When justified to Bottom, the font so that its baseline appears at the
specified position (note that glyphs with descenders may still draw
beneath that point).

These new justification types *do* work with line-wrapped text strings.