Scenes don't get updated, when not enabled.
Also, "FindEntityAtPosition" tests return no values.

(Rendering pipelines already weren't drawing scenes when disabled;
this just adds drawing to the list of things which aren't done on
scenes)

This is a patch on a broken architecture.  Fundamentally, what we
*should* be doing is not having a global set of scenes.  Games should be
responsible for providing a list of scenes to the underlying system,
instead of having them be created by the system and automatically used
universally within a game.

Being able to disable a scene in this way allows us to safely fill
the scene from a background thread, without needing to worry about the
main thread traversing the scene graph while the background thread is
modifying it;  letting a game disable scenes so that we won't
automatically traverse their hierarchies is the "smallest possible
change" which makes it safe to work with background threads.

In the future, I should definitely move to the more flexible "just
provide the set of scenes to the system" approach, or make games manage
their scene objects separately.  The only awkwardness is that doing
either of these things would make the current
"vsEntity::RegisterOnScene(int)" utility function impossible -- all the
code everywhere which uses that would have to be rewritten to use
whatever new system took the place of the current one.

And I'm not ready to do that yet.