Reinstated OpCode_SetMatrix4x4.
(Previously, we only supported OpCode_SetMatrices4x4, which took a
pointer, which meant that the matrix had to exist somewhere in memory
for the duration of the display list's lifetime.)

In practice, this "somewhere in memory" was typically in the
vsRenderQueue which assembled the display list.  But as I'm
experimenting with new rendering structures, I've found some situations
where I want to use temporary vsRenderQueue objects to assemble a
portion of a display list, and when that temporary object goes away (or
is otherwise cleared), the memory holding that matrix also goes away.

So for now, let's just support putting the matrix explicitly into the
display list again.

Also, added a different StartRender() function on the vsRenderQueue, to
make it easier to create temporary vsRenderQueue objects.