Set 'active bytes' when updating VBO contents.
This really should be expected behaviour.  It was already doing this if
you set an external array into a VBO, but wasn't doing it if you were
directly updating the vsRenderBuffer's internal array cache, or if your
vsRenderBuffer didn't have a VBO attached to it.

Now it *always* updates the vsRenderBuffer's 'active bytes', regardless
of those situations.