Switch to a simple multiply-with-carry implementation of random.
(Previously, we used an implementation of Mersenne Twister)

This gives us a number of benefits, the largest of which is that the
code is now simple enough to quickly audit, but also that we now can
very easily have multiple random generators running at once.

This means that we can (for example) safely use random generators from
different threads (one per thread), or having one random generator which
creates reliable random numbers which will work out the same way every
time, and another which is used for unimportant ones.