Starting to segregate old-style rendering.
Old-style push/pop of matrices, applications of color/material, etc. are
no longer put into DisplayLists by default.  We now only do that if an
old-style display list has been set explicitly on a sprite or model.

And even in that case, we now set the full matrix, instead of pushing it
on top of the "old" matrix.

In new-style games, this can make a *huge* difference to the number of
commands going into the rendering queue.  (In MMORPG Tycoon 2, for
example, this reduces our highwater FIFO usage from about 150kb of
rendering commands per frame to about 50kb per frame.)