Adds a display showing current rendering command fifo usage.
This appears immediately beneath the timing bars.  Note that
we currently assert if the rendering command queue every fills
up (and there's really no good 'fix' if it does), so it's important to
keep an eye on it!

TODO is a way for a game to set the size of its rendering command queue.

In MMORPG Tycoon 2, up until the 'new_style' branch, I was using up to
300kb of command queue.  In this branch, it doesn't go much over 30kb
any more.  Which still seems a bit high, but is at least vaguely
reasonable.