Work-in-progress of updated material system.
This new system allows the setting of shader variables on specific
instances of a vsMaterial.

So far, only supported for uniforms, and only for boolean and float
values.  But this is still very early days, and the render command
queue format is likely to change more before this settles.

Net result of this is that you could (for example) set a single "effect"
shader on several different models, but provide different parameters
values to each, even though they're using the same shader and the same
underlying "material" file.

Also, this fixes a minor bug which was causing the vsRenderQueue to not
batch material usage properly (it was batching based on material
instances, not based on the real internal material referenced by those
instances).  Might result in some small performance wins.