Proof-of-concept support for uniform buffers.
With this commit, we now transmit view and projection matrices to
shaders inside a uniform buffer.  Since these two matrices tend to be
relatively stable across a frame (typically only changing when switching
between scenes), this should save a lot of explicit uniform setting.
And it's a first step toward broader, perhaps game-side application of
generic uniform buffers.

Note that due to this change, if you're using custom shaders, you need
to be taking view and projection matrices from the uniform buffer -- we
no longer support explicit, loose uniforms for these, so you'll need
to update your shaders.  Check the diffs on default_v.glsl in this
commit to see the small shader shader code modification which will be
required for custom game shaders.