Adds vsModelInstanceGroup.
This allows you to create several different "groups" of instances, which
can be drawn separately.  The vsModelInstanceGroup can be hooked into a
scene graph, and will draw its own set of instances, when its Draw
function is called.

The old interface on the vsModel is still present;  the vsModel just has
a built-in vsModelInstanceGroup which it passes calls through to.