Utility function for querying text dimensions.
Two methods;  one which creates a full display list and then inspects it
to determine its side, and a maths-based one which tries to figure it
out without creating a full display list.

The maths-based system seems accurate enough to be acceptable, so am
going with that for the moment.