Fix zwrite=off and zread=on
Previously, we were disabling the use of the depth buffer entirely, if
a material's zread was turned off.  Which meant that turning zread off
also implicitly turned off zwrite.

Now we instead set the depth test to always pass, if zread is turned
off, but still will write depth values into the depth buffer if zwrite
is on.

Additionally, if *both* zread and zwrite are off, then we disable use of
the depth buffer entirely.

"Ignore the current depth values, but overwrite the depth buffer
with my values" is an *extremely* unusual thing to want to do.  Which is
why I hadn't noticed the problem until now.  This change should have no
visible effect on anything, apart from that one situation.