Removed old vsRenderBuffer::MapAll and UnmapAll functions.
We used to use these utility functions when we needed to create a new
OpenGL context during a resize event.  However, we haven't needed to
create new contexts at that time for well over a year.  And so MapAll
and UnmapAll are no longer needed.

As a side-benefit, this means that the vsRenderBuffer doesn't need to
have an AutomaticInstanceList, which makes it easier to use in
multithreaded code.

(An alternate fix would have been to make the AutomaticInstanceList
thread-safe.  But in practice, I'd rather simply not use it, since it's
not giving us any benefits.)