POC for "vsShaderValues"; override shader variables for a draw
Lets you set generic uniform variables which will be used in preference
to those set on a material.

Right now, the only access to this is through model instance groups
(since that's where I was wanting to test it).  This is just a general
proof of concept for the interface and the way it might work.  There
will probably be major revisions to this in the near future.