vsMaterial now asserts if you request a material which doesn't exist.
Previously, it'd silently create a default material, which could
theoretically result in one debugging for a few hours to figure out why
a particular material's shader wasn't working correctly, only to
eventually realise that it's because the material file wasn't actually
present on the disk, and the engine hadn't bothered to tell you.

Theoretically.