Re-enable explicit buffers.
Fixed the bug which was causing flicker and etc;  when rendering
instances straight from a VBO, the renderer assumed that the size of
the VBO told us the number of instances.  Which it does.  Except that it
was looking at the *allocated* size, instead of the *active* size.  And
since we just use less of the VBO if we start drawing fewer instances,
that was resulting in all sorts of redraws, and since colors weren't
being updated for those out-of-bounds instances, that was leading to
platform-specific flickering, etc.

All fixed now!