Proof of concept for non-VBO triangle list drawing.
Uses glDrawArrayInstanced instead of glDrawElementsInstanced().

Not sure if this is really a good way to do this.  Maybe a better choice
would be for the vsRenderBuffer to notice if it's a trivial index
buffer, and if so, to just call glDrawArraysInstanced() instead of
glDrawElementsInstanced().  That would be much lower-intrusion into the
rest of the codebase.