Use spin locks instead of mutexes on OS X.
OS X mutexes are very expensive, heavyweight affairs.  In debug builds
on my laptop, doing a malloc often takes several thousand times as long
to lock and unlock the allocator's mutex than it takes to find and
prepare a block of memory.

So let's switch the OS X mutex to a spinlock;  this brings the
lock/unlock time down to almost zero, and makes a huge difference for
Mac builds which use mutexes.