Adds support for grabbing depth textures from a vsRenderTarget.
These textures can then be used in shaders, for whatever purpose you
like.

Caveat:  Currently, all depth textures are now being set to compare_mode
"GL_NONE".  Which means that they can't be sampled as shadow textures.
(Which we didn't really support properly anyhow;  I'll need to think
about that at some point in the future)

Eventual plan is to use this to experiment with an implementation of SSAO.