Correctly calculate cube diagonal for box classification.
Incorrect calculation was causing us to check a sphere which was
sometimes too small to contain the box;  this meant that we sometimes
declared a box as being not-visible when it actually was, on the basis
that a sphere which contained the box wasn't visible.  (but the sphere
we were testing was actually smaller than the box!)

Now fixed.