Initialise file length, even if file fails to open.
Mingw helpfully points out that it's possible that the vsFile::m_length
might not be set, if the requested file couldn't be opened.  Of course,
if that had happened then our vsAssert would have fired and everything
would have stopped.  But it's still a fair point;  vsAsserts could
theoretically be compiled out in release, so we'd be much better off
initialising critical values like m_length, just to be certain that we
don't try to do anything silly with non-existant files.