Add support for disabling SDL_Mixer and Box2D.
If SDL_Mixer is disabled, it will be excluded from VectorStorm builds,
and the whole VS SoundSystem will be entirely excluded.  This allows
games to integrate other sound solutions, instead of using the
simplistic built-in one.

Similarly, if Box2D is disabled, it will be excluded from VectorStorm
builds, and the whole VS Physics/Collision system will be entirely
excluded.  This allows games to not need to distribute the Box2D
libraries if they're not actually using them.

These are controlled using the "USE_SDL_SOUND" and "USE_BOX2D_PHYSICS"
options, accessible through cmake.