VS_OVERRRIDE_ALLOCATOR is now a build option.
Previously, this was a #define that needed to be passed in manually from
somewhere unspecified.  Now it's configured via the VS_Config.h
mechanism, which means that it can be set manually, or better, can be
set from a game's CMakeLists.txt file based on any desired sort of
criteria.