Okay, forcing acceptance of debug callback.
GLEW bug #223 reports that GLEW is misreporting GLDEBUGPROC as requiring
'GLvoid *' as its final parameter, where the official spec says that it
should be 'const GLvoid *'.  This bug was fixed in GLEW 1.11.  Sadly,
on half my machines, I'm using a GLEW version earlier than that;  most
notably, the chroots that Valve provides for making Linux Steam builds
use a GLEW which does not have this fix.

Rather than go in and work around that issue in those chroots, I'm just
going to force the issue by making the compiler accept the callback
function I've provided, even if it doesn't quite match what GLEW
(incorrectly) declares it should be expecting.