Workaround for NVidia Optimus support.
NVidia Optimus doesn't autoswitch to NVidia card for OpenGL contexts;
only for DirectX and OpenCL.  This is a hack which exports a global
variable which *should* be picked up by the Optimus' driver and force
it to switch rendering to the NVidia card.

I don't have access to this hardware for testing, so this is a
proof-of-concept change which I'm going to send out to testers.
Crossing my fingers!