Ugly workaround;  '-nan' gets interpreted as 0.
'nan' already got interpreted as 0 (by being interpreted as a string
without a numeric value), but '-nan' was getting parsed into '-' and
'nan', and '-' wasn't being handled.  This works around the problem
in probably the ugliest way possible, seeing as 'nan' really doesn't
equal zero.  But this at least works around the assertion on line 277.