Fix tromping of "camera is reference" value when drawing with custom camera
Honestly, vsScene shouldn't have a camera set on it.  Cameras and scenes
should be entirely separate, joined only by a rendering pipeline stage.
But that would be a big refactoring at this stage.  I'll do it someday,
but not today.  For now, this fixes the problem of tromping over that
"camera is reference" variable;  we now restore the old value of the
variable after drawing with a custom camera.