Initial revision of vsRenderTarget::Resize support.
This allows changing the width/height of a render target, and all
attached textures.

Note that this is currently only implemented for pure texture
attachments, NOT for render buffers.  So if you have multisampling
enabled, this won't currently work for you.  (It might be easy to
support;  I just haven't needed it yet, so I've left it unimplemented
and uninvestigated)