Fix vsFile::ReadLine() on the last line of the file.
Was previously ignoring the final character, using assuming that the
character was a newline.  Now we actually check, instead of assuming.

Fixes file loading bugs for people who have edited text data files in
notepad, or other text editors which don't enforce the presence of a
newline at the end of the final line of a text file.