Adds GL_CHECK_SCOPED("context") and GL_CHECK("context") macros.
Scoped checks for OpenGL errors both at the start and end of the scope,
and non-scoped checks immediately.

Adds a new option "VS_GL_DEBUG".  If turned on, the above macros
actually check for errors;  if off, they become NOPs so strongly that
their strings won't even make it into the string table.  (Yay for
preprocessor macros!)

Additionally, if that option is enabled, we try to set up an OpenGL
debug context instead of a core context, and if that's successful we
try to hook into the OpenGL debug message callback.  If that's
successful, then any OpenGL debug messages get put into our log as well.