Set up global variable to cause a crash on Texture state warning.
Defaults to disabled.  But a game could tweak the variable to make the
program crash if this texture state warning is sent, and if that game
had set up some sort of way to log out a backtrace and file a bug
report, that might be really useful for somebody who wanted to know
about what the callstack was like when that trace was received.

Just saying.  In case somebody was in that position.

Obviously not me.