Adds vsRenderer::ScreenshotBack(), which captures the back buffer.
(Resolving any MSAA, as required).

Also fixes ::Screenshot() to sample from the default framebuffer, not
the back buffer that we're using targetting.  Haven't done that for
Alpha or Depth screenshots yet, since those values don't ever make it to
the main default framebuffer, but that means that they'll fail if
they're called while MSAA is enabled.  Need to clean that up at some
point.