ResolveRenderTarget op 'NULL' param == main render target.
Matches the meaning of a 'NULL' param for a SetRenderTarget render op.

Not sure this is a good idea;  probably we should be expecting people to
actually tell us to use the main render target if that's what they mean,
but I've already established this as what 'NULL' means, regarding render
targets, so..  might as well stick to it until I've decided to change
the policy.

Fixes issues where the main render target wasn't getting resolved when
people (i.e.: me) expected it to.