Adds a utility function for vsMaterialInternal::SetTexture().
Strictly speaking, games should never do that.  a vsMaterialInternal
should be created in one go and never modified.  And if you want to
set one up or modify it in code, do it through a vsDynamicMaterial.

...but in practice, I needed to do it, and if you need to do it, you
need to do it correctly or you wind up with memory leaks or other
problems, so I've provided a function that does it safely/correctly.