Adds a vsTweenSpline3D as a proof of concept.
I was trying to make a vsTween<> object which would work nicely for
continuously-moving objects, and so tried this.  However, the real
problem is that this sort of approach winds up lagging behind the moving
object, which wasn't what I wanted for my specific task today.  However,
this object could still be useful someday, so I'm committing it, even
though I'm not actually using it right now.