Various Font rendering fixes.
1. Fixes incorrect ignore of font line spacing.

(It was trying to round a value down to integer pixel values, when it
wasn't in pixel units.  So extra space between lines would be rounded
instead to an integer number of line heights.  Now works correctly, and
only rounds to pixels at the last minute, before emitting glyphs.)

2. Fixes GetStringDimensions() to get the width only of the widest
wrapped line, not of the whole unwrapped string.

3. Fixes vsFontSize::GetStringWidth() to use the whole width of the
last character of a string.  (Previously, it only used the xAdvance
for that character, which meant that it would often ignore the last
half of the last character in a line, if the cursor didn't advance
all the way past the end of the glyph).  This provides *far* more
accurate results.