Use smaller font sizes, when text is shrunk to fit.
Previously, if you were drawing a string using a 32-point font and it
got shrunk down to 16-point in order to fit into its specified bounds,
we would shrink down the 32-point version of the font.

Now, we draw using a 16-point version of the font, if available.  (We
use the smallest size version of the font which is >= the size we shrunk
to).

Also, when shrinking to try to fit a string into a box, we now step down
by 1.0f pixels at a time.  Previously, we shrunk by 5% at a time.