By default. vsFile now writes into a temporary file.
vsFile moves its written file into position once the vsFile is destroyed
(and therefore writing is finished).  This means that you don't wind up
with partial files sitting on the disk in the event of a crash.
Temporary files which are currently being written into are placed inside
a 'tmp' directory in the write directory, in a format which matches the
non-tmp directory structure.

If you have a use case which really requires you to write out to a file
and have that file actually exist in its final position while it's being
written into, you can use vsFile's new mode "MODE_WriteDirectly" instead
of "MODE_Write".  But in the vast majority of cases, you can completely
ignore this change and simply enjoy the lack of file corruption.