Remove cache of 'next' entity during ::Update()
It was possible for an entity's Update() to destroy the next object that
happened to be in the queue.  And that could send the software into an
infinite loop.

Now, it's actually just straight-up illegal to delete yourself during
your own ::Update() call, instead of me trying to be clever and making
the software try to cope with it.