Workaround support for "queued extract" operation.
This is NOT an okay solve;  it relies on objects not overriding their
"Update()" and "Draw()" calls.  Which obviously everything does.

Maybe this needs another, non-virtual "Update"-like method which then
invokes the virtual "Update"?  Or something like that?  Just in order to
keep track of whether we're in a dangerous bit of code or not.

But for right now, this gets around the immediate issue.