Array constructors made explicit.
If you write this code:

    vsArray<foo> myArray = 2;

Previously, the '2' would have been auto-converted into a vsArray with
pre-allocated space for two elements.  Clearly, that is surprising
nonsense.

We now we get a compiler error, instead.

If you really wanted to replace the array with a new one, do that
explicitly, like this:

    vsArray<foo> myArray = vsArray<foo>(2);