First round of input refactoring.
The default set of control inputs are now configured in code.  Still
TODO is to add a build configuration option which makes us NOT make
those default inputs, save out the current set of bindings to the
game-defined axes, and provide a re-binding interface.

Plus re-implementing support for game controllers.  (There actually
isn't a whole lot of work that needs to be done around that)