Add build setting to disable default controller configuration.
vsSimpleMenu is removed if you do this, since it doesn't know what
control codes to use.

Also did some fiddling with memory management in vsInput, since adding
axes and bindings can allocate memory which doesn't get deallocated on
destruction, so that now gets allocated out of the global heap instead
of the game heap.  If you're using overloaded allocators which care
about memory leaks, that is.