Support for physfs v2.1+
In physfs v2.1.0, a whole lot of interfaces got deprecated, and new ones
appeared to take their place.  These were all (I believe) positive
changes.  So I wanted to update the code to make use of the new
functions.  However, most of my build servers are still running on
PhysFS 2.0.3, which didn't yet have the new interfaces.

This commit changes the code to use the new PhysFS 2.1.0 interfaces, and
adds glue code providing those interfaces where they're missing in
earlier verisons of the PhysFS library.  This means that it should be
easy to just drop in new versions of PhysFS, without requiring any
VectorStorm changes in the future.