Experimentally adding a const operator[] to vsArrayStore.
C++ compilers can be touchy about providing both a const and non-const
operator[], and intelligently selecting which one to call.  But.. it
seems to work so far.  We'll see what happens when the build machine
tries to use this stuff using the SteamOS build system with its older
compilers.  Ready to back this change back out, if necessary.  But it's
a nice Quality Of Life thing, if I can keep it.