vsCache and vsHashTable use a Fibonacci hash, instead of modulo.
Previously, we used integer modulo to convert from a string hash value
to a bucket index.  This is slow, because it involves integer division.

Now, we instead use a Fibonacci hash, which uses just an integer
multiply and a shift.  This is *heaps* faster.

Of course, it only matters if you're doing a lot of hash/cache lookups.
In the case of MMORPG Tycoon 2, making this change reduced the time
taken to load a saved game (on my computer) from 5.5 seconds to 5.3
seconds.  Not a huge win, but not too bad for such a small change!