Provides a variable containing include paths.
We had an issue where VectorStorm needed to provide access to
VS_Config.h settings for client code, but that header is generated
during the build, and is in a surprising place during the build process.

As a result, we now provide a variable, `VECTORSTORM_INCLUDE_PATHS`,
which projects using VectorStorm should add to their own include paths.